update schedule error for different languages
