+++ Tracker/Extensions/DateFormatter+weekday Extension.swift
@@ -9,18 +9,36 @@ import Foundation
 
 extension DateFormatter {
     
+//    static func weekday(date: Date) -> String {
+//        let weekdayNumber = Calendar.current.dateComponents([.weekday], from: date).weekday
+//        var weekday = "Нет такого дня недели"
+//        if let weekdayNumber = weekdayNumber {
+//            switch weekdayNumber {
+//            case 1: weekday = Weekdays.Sunday.rawValue
+//            case 2: weekday = Weekdays.Monday.rawValue
+//            case 3: weekday = Weekdays.Tuesday.rawValue
+//            case 4: weekday = Weekdays.Wednesday.rawValue
+//            case 5: weekday = Weekdays.Thursday.rawValue
+//            case 6: weekday = Weekdays.Friday.rawValue
+//            case 7: weekday = Weekdays.Saturday.rawValue
+//            default: weekday = "Нет такого дня недели"
+//            }
+//        }
+//        return weekday
+//    }
+    
     static func weekday(date: Date) -> String {
         let weekdayNumber = Calendar.current.dateComponents([.weekday], from: date).weekday
         var weekday = "Нет такого дня недели"
         if let weekdayNumber = weekdayNumber {
             switch weekdayNumber {
-            case 1: weekday = Weekdays.Sunday.rawValue
-            case 2: weekday = Weekdays.Monday.rawValue
-            case 3: weekday = Weekdays.Tuesday.rawValue
-            case 4: weekday = Weekdays.Wednesday.rawValue
-            case 5: weekday = Weekdays.Thursday.rawValue
-            case 6: weekday = Weekdays.Friday.rawValue
-            case 7: weekday = Weekdays.Saturday.rawValue
+            case 1: weekday = "7"
+            case 2: weekday = "1"
+            case 3: weekday = "2"
+            case 4: weekday = "3"
+            case 5: weekday = "4"
+            case 6: weekday = "5"
+            case 7: weekday = "6"
             default: weekday = "Нет такого дня недели"
             }
         }
